+++ InterneTaakAfhandeling.Web.Client/src/views/DashboardView.vue
@@ -13,7 +13,11 @@
 
     <section>
       <utrecht-heading :level="2" id="h2-b">Oudste contacten voor afdeling</utrecht-heading>
-      <interne-taak-table v-if="!isLoading" :interne-taken="staticRecords" aria-labelledby="h2-b" />
+      <interne-taak-table
+        v-if="!isLoading"
+        :interne-taken="fakeInterneTaken"
+        aria-labelledby="h2-b"
+      />
     </section>
   </div>
 </template>
@@ -33,32 +37,11 @@
 <script setup lang="ts">
 import { onMounted } from "vue";
 import { useUserStore } from "@/stores/user";
-import { storeToRefs, type _DeepPartial } from "pinia";
+import { storeToRefs } from "pinia";
 import InterneTaakTable from "@/components/InterneTaakTable.vue";
-import type { Internetaken } from "@/types/internetaken";
+import { fakeInterneTaken } from "@/helpers/fake-data";
 const userStore = useUserStore();
 const { assignedInternetaken, isLoading } = storeToRefs(userStore);
-
-const _staticRecords: _DeepPartial<Internetaken>[] = [
-  {
-    uuid: "1",
-    aanleidinggevendKlantcontact: {
-      plaatsgevondenOp: "2025-01-01T15:00+01:00",
-      onderwerp: "Lorem ipsum"
-    },
-    betrokkene: { volledigeNaam: "Saskia Swart" }
-  },
-  {
-    uuid: "2",
-    aanleidinggevendKlantcontact: {
-      plaatsgevondenOp: "2025-01-02T15:00+01:00",
-      onderwerp: "Dolor sit amet"
-    },
-    betrokkene: { volledigeNaam: "Martijn de Groot" }
-  }
-];
-const staticRecords = _staticRecords as unknown as Internetaken[];
-
 onMounted(() => {
   userStore.fetchAssignedInternetaken();
 });
